+++ retrofit/src/main/java/retrofit/http/mime/TypedByteArray.java
@@ -1,5 +1,5 @@
 // Copyright 2010 Square, Inc.
-package retrofit.io;
+package retrofit.http.mime;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
