+++ retrofit/src/test/java/retrofit/http/client/ApacheClientTest.java
@@ -17,8 +17,8 @@ import org.apache.http.message.BasicHttpResponse;
 import org.apache.http.message.BasicStatusLine;
 import org.junit.Test;
 import retrofit.http.Header;
-import retrofit.io.TypedOutput;
-import retrofit.io.TypedString;
+import retrofit.http.mime.TypedOutput;
+import retrofit.http.mime.TypedString;
 
 import static org.fest.assertions.api.Assertions.assertThat;
